# Leaf 4381 - Optimize formGrid.js rendering