--- a/LEAF_Request_Portal/js/formQuery.js
+++ b/LEAF_Request_Portal/js/formQuery.js
@@ -302,7 +302,7 @@ var LeafFormQuery = function () {
     if (query.getData != undefined && query.getData.length == 0) {
       delete query.getData;
     }
-    if (query.limit == undefined || isNaN(query.limit) || parseInt(query.limit) > 1000) {
+    if (query.limit == undefined || isNaN(query.limit) || parseInt(query.limit) > 1000 || !isFinite(query.limit)) {
       return getBulkData();
     }