--- a/LEAF_Request_Portal/js/formQuery.js
+++ b/LEAF_Request_Portal/js/formQuery.js
@@ -331,6 +331,7 @@ var LeafFormQuery = function () {
     updateTerm,
     updateDataTerm,
     setQuery: (inc) => query = inc,
+    setBatchSize: (size) => batchSize = size,
     setLimit,
     setLimitOffset,
     setRootURL: (url) => rootURL = url,