--- a/LEAF_Request_Portal/js/formQuery.js
+++ b/LEAF_Request_Portal/js/formQuery.js
@@ -343,6 +343,7 @@ var LeafFormQuery = function () {
     onSuccess,
     onProgress,
     setAbortSignal,
+    getResults: () => results,
     execute
   };
 };