@mgaoVA 2024-06-04
Flagging as draft, issue with needing to re-apply events after rendering from cache