@mgaoVA 2024-06-05
@aerinkayne The issue with events should be fixed now. Used a different caching approach, and I also updated the test page to include event listeners